merge(bento-get): updates
