+++ code/packages/bento-get/src/hooks/useFetchComponent.tsx
@@ -3,8 +3,9 @@ import querystring from 'node:querystring'
 import { Octokit } from 'octokit'
 import React from 'react'
 import useSWR from 'swr'
-import { AppContext, debugLog } from '../commands/index.js'
 import { useNavigate } from 'react-router-dom'
+import { AppContext } from '../data/AppContext.js'
+import { debugLog } from '../commands/index.js'
 
 interface GithubUserData {
   login: string
