+++ code/packages/bento-get/types/screens/CodeAuthScreen.d.ts
@@ -1,0 +1,2 @@
+export declare const CodeAuthScreen: () => import("react/jsx-runtime").JSX.Element;
+//# sourceMappingURL=CodeAuthScreen.d.ts.map
